Stylistic and copy edits to failed upload email
I made some stylistic and copy edits to this email template, giving it a slightly more conversational tone.

A question: What action should the user take after reading this email? (My guess is that they should reach out to the person who tried to upload data & ask them to try again, but this may not be correct.) It could be helpful to include a clear call to action so the user knows what to do next. (That said, if the user already knows what to do next, we may not need a CTA.)