V. quick copy edits
Just added a salutation/closing. And a question for you all — do we need to include a CTA here, or will recipients know what to do? Follow-up: Are there any tips/strategies we can offer to help folks from having their price lists rejected? If a tip sheet doesn't exist (/if it's something you're interested in), I'd be happy to draft one ...